+++ src/mserv/cli/cli.py
@@ -1,5 +1,7 @@
 import argparse
 import logging
+import queue
+
 from mserv.services.samples import MessageGenerator, MessageLogger
 from mserv.socketlib import (
     Client,
@@ -17,6 +19,8 @@ def start_socket(
         client: bool,
         sock_type: str,
         reconnect: bool,
+        timeout: float,
+        messages: list[str],
         logger: logging.Logger
 ) -> None:
     valid_types = ["multi", "receiver", "sender"]
@@ -34,25 +38,78 @@ def start_socket(
 
     if sock_type == "multi":
         if client:
-            socket = Client(address, reconnect=reconnect, logger=logger)
+            socket = Client(
+                address,
+                reconnect=reconnect,
+                timeout=timeout,
+                logger=logger
+            )
         else:
-            socket = Server(address, reconnect=reconnect, logger=logger)
+            socket = Server(
+                address,
+                reconnect=reconnect,
+                timeout=timeout,
+                logger=logger
+            )
         msg_logger = MessageLogger(socket.received, logger)
         msg_gen = MessageGenerator(socket.to_send, name=name, logger=logger)
 
     elif sock_type == "receiver":
         if client:
-            socket = ClientReceiver(address, reconnect=reconnect, logger=logger)
+            socket = ClientReceiver(
+                address,
+                reconnect=reconnect,
+                timeout=timeout,
+                logger=logger
+            )
         else:
-            socket = ServerReceiver(address, reconnect=reconnect, logger=logger)
+            socket = ServerReceiver(
+                address,
+                reconnect=reconnect,
+                timeout=timeout,
+                logger=logger
+            )
         msg_logger = MessageLogger(socket.received, logger)
 
     elif sock_type == "sender":
-        if client:
-            socket = ClientSender(address, reconnect=reconnect, logger=logger)
+        if not messages:
+            if client:
+                socket = ClientSender(
+                    address,
+                    reconnect=reconnect,
+                    timeout=timeout,
+                    logger=logger
+                )
+            else:
+                socket = ServerSender(
+                    address,
+                    reconnect=reconnect,
+                    timeout=timeout,
+                    logger=logger
+                )
+            msg_gen = MessageGenerator(socket.to_send, name=name, logger=logger)
         else:
-            socket = ServerSender(address, reconnect=reconnect, logger=logger)
-        msg_gen = MessageGenerator(socket.to_send, name=name, logger=logger)
+            to_send = queue.Queue()
+            for msg in messages:
+                to_send.put(msg)
+            if client:
+                socket = ClientSender(
+                    address,
+                    to_send=to_send,
+                    reconnect=False,
+                    timeout=timeout,
+                    stop=lambda: to_send.empty(),
+                    logger=logger
+                )
+            else:
+                socket = ServerSender(
+                    address,
+                    to_send=to_send,
+                    reconnect=False,
+                    timeout=timeout,
+                    stop=lambda: to_send.empty(),
+                    logger=logger
+                )
 
     else:
         raise ValueError(f"Unexpected type {sock_type}")
@@ -120,22 +177,39 @@ def parse_args():
         "--reconnect",
         "-r",
         action="store_true",
-        help="Whether the client or server should try to reconnect if the connection is lost"
+        help="Whether the client or server should try to reconnect if the connection is lost."
+    )
+    parser.add_argument(
+        "--timeout",
+        "-o",
+        type=float,
+        default=5,
+        help="Timeout for socket receive and send operations."
+    )
+    parser.add_argument(
+        "--messages",
+        "-m",
+        type=str,
+        nargs="+",
+        help="A set of messages that will be sent by a client sender or server sender."
+             " The program will exit after sending all messages. "
     )
 
     args = parser.parse_args()
     address = (args.ip, args.port)
-    return address, args.server, args.type, args.reconnect
+    return address, args.server, args.type, args.reconnect, args.timeout, args.messages
 
 
 def main():
-    address, server, sock_type, reconnect = parse_args()
+    address, server, sock_type, reconnect, timeout, messages = parse_args()
     logger = get_module_logger(__name__, config="dev", use_file_handler=False)
     start_socket(
         address,
         client=not server,
         sock_type=sock_type,
         reconnect=reconnect,
+        timeout=timeout,
+        messages=messages,
         logger=logger
     )